requirements.txt
flask==3.0.2
flask-cors==4.0.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
google-genai>=0.7.0
requests==2.31.0
gunicorn==21.2.0


